refactor PortfolioStatusExcelTableFactory
